+++ spec/ci_config_spec.rb
@@ -110,7 +110,7 @@ RSpec.describe ArduinoCI::CIConfig do
 
   context "allowable_unittest_files" do
     cpp_lib_path = File.join(File.dirname(__FILE__), "fake_library")
-    cpp_library = ArduinoCI::CppLibrary.new(cpp_lib_path, "my_fake_arduino_lib_dir")
+    cpp_library = ArduinoCI::CppLibrary.new(cpp_lib_path)
 
     it "starts with a known set of files" do
       expect(cpp_library.test_files.map { |f| File.basename(f) }).to match_array([